link newAdModal to header
