+++ frontend/src/pages/Dashboard.tsx
@@ -1,36 +1,17 @@
-import { useState } from "react";
 import { useAuth } from "../contexts/AuthContext";
 import { Box, Button } from "@mui/material";
-import GenericModal from "../components/Modal/GenericModal";
-import AdModalForm from "../components/NewAdModal/AdModalForm";
 import DashboardSection from "../components/Dashboard/DashboardSection";
 
 export default function Dashboard() {
-  // To delete once header is available */
+  // To delete once header is finalised
   const { logout } = useAuth();
-  const [newAdModalIsOpen, setNewAdModalIsOpen] = useState<boolean>(false);
-  const closeNewAdModal = () => {
-    setNewAdModalIsOpen(false);
-  };
 
   return (
     <>
-      {/* To delete once header is available */}
+      {/* To delete once header is finalised */}
       <button onClick={() => logout()}>Se déconnecter</button>
-      <Button onClick={() => setNewAdModalIsOpen(true)}>
-        Créer une annonce
-      </Button>
-      {newAdModalIsOpen && (
-        <GenericModal
-          open={newAdModalIsOpen}
-          onClose={closeNewAdModal}
-          maxWidth="md"
-        >
-          <AdModalForm onClose={closeNewAdModal} />
-        </GenericModal>
-      )}
 
-      {/* Real Dashboard */}
+      {/*  Dashboard */}
       <img
         src="/images/dashboard-poster.png"
         alt="helping to walk"
